@@ -19,29 +19,28 @@
   },
   "private": true,
   "devDependencies": {
-    "@angular/animations": "7.2.15",
-    "@angular-builders/custom-webpack": "7.3.1",
-    "@angular/cli": "7.3.9",
-    "@angular/common": "7.2.15",
-    "@angular/compiler": "7.2.15",
-    "@angular/compiler-cli": "7.2.15",
-    "@angular/core": "7.2.15",
-    "@angular-devkit/build-angular": "0.13.9",
-    "@angular-devkit/build-optimizer": "0.13.9",
-    "@angular-devkit/core": "7.3.9",
-    "@angular/forms": "7.2.15",
-    "@angular/http": "7.2.15",
-    "@angular/language-service": "7.2.15",
-    "@angular/platform-browser": "7.2.15",
-    "@angular/platform-browser-dynamic": "7.2.15",
-    "@angular/router": "7.2.15",
-    "@ngtools/webpack": "7.3.9",
-    "@ngrx/effects": "7.4.0",
-    "@ngrx/router-store": "7.4.0",
-    "@ngrx/store": "7.4.0",
-    "@ngrx/store-devtools": "7.4.0",
+    "@angular/animations": "8.2.2",
+    "@angular-builders/custom-webpack": "8.1.0",
+    "@angular/cli": "8.2.1",
+    "@angular/common": "8.2.2",
+    "@angular/compiler": "8.2.2",
+    "@angular/compiler-cli": "8.2.2",
+    "@angular/core": "8.2.2",
+    "@angular-devkit/build-angular": "0.802.1",
+    "@angular-devkit/build-optimizer": "0.802.1",
+    "@angular-devkit/core": "8.2.1",
+    "@angular/forms": "8.2.2",
+    "@angular/language-service": "8.2.2",
+    "@angular/platform-browser": "8.2.2",
+    "@angular/platform-browser-dynamic": "8.2.2",
+    "@angular/router": "8.2.2",
+    "@ngtools/webpack": "8.2.1",
+    "@ngrx/effects": "8.2.0",
+    "@ngrx/router-store": "8.2.0",
+    "@ngrx/store": "8.2.0",
+    "@ngrx/store-devtools": "8.2.0",
     "@ngrx/store-log-monitor": "3.0.2",
-    "angular-confirmation-popover": "4.2.0",
+    "angular-confirmation-popover": "4.2.1",
     "angular-notifier": "4.1.1",
     "bootstrap": "3.4.1",
     "codelyzer": "4.5.0",
@@ -56,19 +55,18 @@
     "karma-jasmine": "1.1.2",
     "karma-jasmine-html-reporter": "1.3.1",
     "ng2-dragula": "2.1.1",
-    "ng2-nouislider": "1.7.13",
-    "ngx-clipboard": "10.0.0",
-    "ngx-select-ex": "3.6.8",
-    "ngrx-store-freeze": "0.1.9",
-    "nouislider": "13.1.1",
+    "ng2-nouislider": "1.8.2",
+    "ngx-clipboard": "12.2.0",
+    "ngx-select-ex": "3.7.0",
+    "ngrx-store-freeze": "0.2.4",
+    "nouislider": "14.0.2",
     "protractor": "5.4.0",
-    "reselect": "3.0.1",
-    "rxjs": "6.4.0",
+    "rxjs": "6.5.2",
     "semver": "5.6.0",
     "ts-keycode-enum": "^1.0.6",
     "uhk-common": "1.0.0",
     "xml-loader": "1.2.1",
-    "zone.js": "0.8.26"
+    "zone.js": "0.10.0"
   },
   "dependencies": {
     "classlist.js": "1.1.20150312",
