@@ -21,7 +21,7 @@ export class SvgTextIconKeyComponent implements OnChanges {
     useY: number;
     textY: number;
     textAnchor: string;
-    spanX: number;
+    textX: number;
     secondaryTextY: number;
     secondaryHeight: number;
 
@@ -46,7 +46,7 @@ export class SvgTextIconKeyComponent implements OnChanges {
         this.useY = (this.width > 2 * this.height) ? this.height / 3 : this.height / 2;
         this.textY = ((this.width > 2 * this.height) ? this.height / 2 : this.height / 3) - textYModifier;
         this.textAnchor = (this.width > 2 * this.height) ? 'end' : 'middle';
-        this.spanX = (this.width > 2 * this.height) ? 0.6 * this.width : this.width / 2;
+        this.textX = (this.width > 2 * this.height) ? 0.6 * this.width : this.width / 2;
 
         this.secondaryHeight = this.height / 4;
         this.secondaryTextY = this.height - this.secondaryHeight - SECONDARY_ROLE_BOTTOM_MARGIN - secondaryYModifier;
